ai_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pgvector>=0.3.0
